Move some weapon code from unit_generic to armed class.
